# changefeedccl: Improve avro encoder performance