@miretskiy 2021-04-19
@ajwerner I realized that after reverting an accidental merge, I never re-reverted this performance improvement PR.